mappers.kt 분리 및 Flow 적용
