+++ app/src/main/java/com/heyday7/movieapp/data/repository/MovieRepository.kt
@@ -4,12 +4,13 @@ package com.heyday7.movieapp.data.repository
 import com.heyday7.movieapp.model.Movie
 import com.heyday7.movieapp.model.Pagination
 import com.heyday7.movieapp.model.SimpleMovie
+import kotlinx.coroutines.flow.Flow
 
 interface MovieRepository {
     suspend fun getMovie(movieId: Int): Movie
-    suspend fun getMovieNowPlaying(
+    fun getMovieNowPlaying(
         page: Int,
         region:String = "KR",
         language: String = "ko"
-    ): Pagination<SimpleMovie>
+    ): Flow<Pagination<SimpleMovie>>
 }
